+++ app/src/cli/new_ember.rs
@@ -5,54 +5,54 @@ use ambient_native_std::git_revision;
 use anyhow::Context;
 use convert_case::Casing;
 
-use super::ProjectPath;
+use super::EmberPath;
 
 pub(crate) fn handle(
-    project_path: &ProjectPath,
+    ember_path: &EmberPath,
     name: Option<&str>,
     api_path: Option<&str>,
 ) -> anyhow::Result<()> {
-    let Some(project_path) = &project_path.fs_path else {
-        anyhow::bail!("Cannot create project in a remote directory.");
+    let Some(ember_path) = &ember_path.fs_path else {
+        anyhow::bail!("Cannot create ember in a remote directory.");
     };
 
     // Build the identifier.
-    let project_path = if let Some(name) = name {
-        project_path.join(name)
+    let ember_path = if let Some(name) = name {
+        ember_path.join(name)
     } else {
-        project_path.to_owned()
+        ember_path.to_owned()
     };
 
-    let name = project_path
+    let name = ember_path
         .file_name()
         .and_then(|s| s.to_str())
-        .context("project path has no terminating segment")?;
+        .context("ember path has no terminating segment")?;
 
-    if project_path.is_dir() && std::fs::read_dir(&project_path)?.next().is_some() {
-        anyhow::bail!("project path {project_path:?} is not empty");
+    if ember_path.is_dir() && std::fs::read_dir(&ember_path)?.next().is_some() {
+        anyhow::bail!("ember path {ember_path:?} is not empty");
     }
 
     let id = name.to_case(convert_case::Case::Snake);
     let id = SnakeCaseIdentifier::new(&id).map_err(anyhow::Error::msg)?;
 
     // Create the folders.
-    let dot_cargo_path = project_path.join(".cargo");
-    let dot_vscode_path = project_path.join(".vscode");
-    let src_path = project_path.join("src");
+    let dot_cargo_path = ember_path.join(".cargo");
+    let dot_vscode_path = ember_path.join(".vscode");
+    let src_path = ember_path.join("src");
 
-    std::fs::create_dir_all(&project_path).context("Failed to create project directory")?;
+    std::fs::create_dir_all(&ember_path).context("Failed to create ember directory")?;
     std::fs::create_dir_all(&dot_cargo_path).context("Failed to create .cargo directory")?;
     std::fs::create_dir_all(&dot_vscode_path).context("Failed to create .vscode directory")?;
     std::fs::create_dir_all(&src_path).context("Failed to create src directory")?;
 
     // Write the files to disk.
-    let ambient_toml = include_str!("new_project_template/ambient.toml")
+    let ambient_toml = include_str!("new_ember_template/ambient.toml")
         .replace("{{id}}", id.as_str())
         .replace("{{name}}", name);
 
     let cargo_toml = {
         // Special-case creating an example in guest/rust/examples so that it "Just Works".
-        let segments = project_path.iter().collect::<Vec<_>>();
+        let segments = ember_path.iter().collect::<Vec<_>>();
         let (replacement, in_ambient_examples) = match segments
             .windows(3)
             .position(|w| w == ["guest", "rust", "examples"])
@@ -88,7 +88,7 @@ pub(crate) fn handle(
             ),
         };
 
-        let template_cargo_toml = include_str!("new_project_template/Cargo.toml");
+        let template_cargo_toml = include_str!("new_ember_template/Cargo.toml");
         let mut template_cargo_toml = template_cargo_toml.replace("{{id}}", id.as_str()).replace(
             "ambient_api = { path = \"../../../../guest/rust/api\" }",
             &replacement,
@@ -108,7 +108,7 @@ pub(crate) fn handle(
         ($path:expr) => {
             (
                 Path::new($path),
-                include_str!(concat!("new_project_template/", $path)),
+                include_str!(concat!("new_ember_template/", $path)),
             )
         };
     }
@@ -131,11 +131,11 @@ pub(crate) fn handle(
     ];
 
     for (filename, contents) in files_to_write {
-        std::fs::write(project_path.join(filename), contents)
+        std::fs::write(ember_path.join(filename), contents)
             .with_context(|| format!("failed to create {filename:?}"))?;
     }
 
-    log::info!("Project \"{name}\" with id `{id}` created at {project_path:?}");
+    log::info!("Ember \"{name}\" with id `{id}` created at {ember_path:?}");
 
     Ok(())
 }