+++ app/src/server/mod.rs
@@ -6,7 +6,7 @@ use std::{
     sync::{Arc, Mutex},
 };
 
-use ambient_core::{asset_cache, name, no_sync, project_name, FIXED_SERVER_TICK_TIME};
+use ambient_core::{asset_cache, ember_name, name, no_sync, FIXED_SERVER_TICK_TIME};
 use ambient_ecs::{
     dont_store, world_events, ComponentDesc, Entity, Networked, SystemGroup, World,
     WorldEventsSystem, WorldStreamCompEvent,
@@ -129,7 +129,7 @@ pub async fn start(
             )
             .unwrap();
 
-        // Keep track of the project name
+        // Keep track of the ember name
         let name = manifest
             .ember
             .name
@@ -138,7 +138,7 @@ pub async fn start(
         server_world
             .add_components(
                 server_world.resource_entity(),
-                Entity::new().with(project_name(), name),
+                Entity::new().with(ember_name(), name),
             )
             .unwrap();